+++ CHANGELOG.md
@@ -4,15 +4,23 @@
 
 ## What's new?
 
-uInt128 struct that works like a regular unsigned integer type.
-
-Int128 and uInt128 explicit conversion operators.
+Int256 struct that works like a regular unsigned integer type.
 
 ---
 
 ## Notes
 
-### Version 0.2.1 - 15/6/2026
+### Version 0.3.0 - 19/6/2026
+
+Add
+
+- Int256 struct. Arithmetical, equality and bitwise operators for the struct. Conversion functions.
+
+Fix
+
+- uInt128 & Int128 ToString conversion methods.
+
+### Version 0.2.1 - 16/6/2026
 
 Add